telegram-service/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
click==8.3.1
fastapi==0.135.2
h11==0.16.0
idna==3.11
pyaes==1.6.1
pyasn1==0.6.3
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
rsa==4.9.1
starlette==1.0.0
Telethon==1.42.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0
